Move ActivityStore into android package.
